Handle missing config file for 11.x inventory
Since we're generating an 11.x inventory before doing the whole 11.x
upgrade, we need to get an 11.x configuration file in place to generate
an accurate inventory. That also needs to be cleaned up so as not to
interfere with the osad upgrade script.

(cherry picked from commit ea2ce8113940a6fc7547b9fb965aa6829f024f64)